aitest/requirements.txt
playwright==1.58.0
pytest==9.0.3
allure-pytest==2.15.3
PyYAML==6.0.3
zhipuai==2.1.5.20250825